Added suport for Signposting relation "identifier"
Of course, if the (URL to the) DOI in the link is one that we cannot parse, the parsing of the initial DOI will fail.

Note: Currently, only one relation is used from the Link header. So if an URL returns a Link header that has multiple URLs having one of the supported relations (linkset, item, identifier), only the first one is considered, the rest are ignored.